refactor: Extract logic from App to BackgroundContainer
Background container is responsible for fetching the background
information. No need to put it in `App`. Save a few rerender.

Deleted HOC since we can use the hook directly.
